Add check for openvpn proccess in htbtarget()
In the htbtarget() function I added a command that checks if a openvpn
procces is running. If that command returns unsuccessful, which would be
a one(1), that means openvpn is not running. There is no reason to execute
rest of the function so it exits the function  by returning 1 for a unsuccessful run.