Fixed VPN session not being created
The error happened when I was updated the if statement to use the $?
bash variable in the previous commit. The error was in the path I typed
for the vpn file. Instead of it being $HOME/ it was $/HOME/ Which broke
file path and didn't allow openvpn the create a session.