Add htbinitfile() func to save file path user want for connection
I added a function which looks for one argument if it is passed in it
check the argument to be a valid file path. If so it then saves the file
path by writing it to a file in it's running directory called `.init`.
When it does not recieve an argument it check to make sure the `.inti`
file exists and then just prints it's contents. When it does not exists
it will warn the user to set a file path.