Have htbinit() use file path in .init file to connect to vpn
I added this by replacing the hard coded path with a substiution where
the path used is from the .init file contents.

I believed this would make the overall function more modular since it
gives us a way to change the path being used for the vpn connection without
having us to provided it everytime or to have to change the source code.